Correction of wrong memory vector scaling and wrong type declaration, introduction of further memory vector scaling
Corrected an error in the resizing of the memory vector, which would base the scaled memory upon the RCEM code string.
Corrected an error in the parsing of the ``z'' commands, which assumed a wrong type declaration for the enumeration of the bits of the I-Cell.
The resizing of the memory vector has been implemented for other commands necessitating such.